Add folders and modules for Api
