+++ app/controllers/api/v1/books_controller.rb
@@ -1,0 +1,33 @@
+module Api
+  module V1
+    class BooksController < ApplicationController
+    
+      def index
+        render json: Book.all
+      end
+    
+      def create
+        book = Book.new(book_params)
+    
+        if book.save
+          render json: book, status: :created
+        else
+          render json: book.errors, status: :unprocessable_entity
+        end
+      end
+    
+      def destroy
+        Book.find(params[:id]).destroy!
+    
+        head :no_content
+      end
+    
+      private 
+    
+      def book_params
+        params.require(:book).permit(:title, :author, :genre, :current_page, :total_pages)
+      end
+    end
+  end
+end
+
